nyanCat refct

+++ src/entities/Entity.java
@@ -3,10 +3,10 @@ package entities;
 import java.awt.*;
 
 public abstract class Entity {
-    public int xPos, yPos, height, width;
+    public double xPos, yPos, height, width;
     public double xDir,yDir;
 
-    public Entity(int xPos, int yPos, int width, int height) {
+    public Entity(double xPos, double yPos, double width, double height) {
         this.xPos = xPos;
         this.yPos = yPos;
         this.width = width;
@@ -15,7 +15,7 @@ public abstract class Entity {
 
     public void render(Graphics g) {
         g.setColor(Color.WHITE);
-        g.fillRect(xPos, yPos, width, height);
+        g.fillRect((int)xPos, (int)yPos, (int)width, (int)height);
     }
 
     public double getYDir() {
